Separate tests from code
