--- src/array.test.js
@@ -1,5 +1,0 @@
-const analyze = require('./array');
-
-it('returns array'), () => expect(analyze([2, 5]).toBe({
-    average: 3.5, min:2, max:5, length:2
-});
