+++ src/code/cipher.js
