+++ src/tests/capitalise.test.js
@@ -1,3 +1,3 @@
-const capitalize = require('./capitalise');
+const capitalize = require('../code/capitalise');
 
 it('capitalize', () => expect(capitalize('microverse')).toBe('Microverse'));
