+++ src/tests/reverseString.test.js
@@ -1,3 +1,3 @@
-const reverseString = require('./reverseString');
+const reverseString = require('../code/reverseString');
 
 it('Reverse given string', () => expect(reverseString('microverse')).toBe('esrevorcim'));
